+++ examples/Connect-AvptElements.ps1
@@ -2,7 +2,7 @@ $clientSecret = ConvertTo-SecureString 'your-client-secret' -AsPlainText -Force
 $credential = [pscredential]::new('your-client-id', $clientSecret)
 
 Connect-AvptElements `
-    -Environment Production `
+    -Environment Commercial `
     -Credential $credential `
-    -TenantName 'contoso-demo'
-
+    -TenantName 'contoso-demo' `
+    -Scope 'your.scope.value'