+++ src/AvePoint.Elements/Private/ConvertTo-AvptPlainText.ps1
@@ -1,0 +1,11 @@
+function ConvertTo-AvptPlainText {
+    [CmdletBinding()]
+    param(
+        [Parameter(Mandatory)]
+        [Security.SecureString] $SecureString
+    )
+
+    $credential = [pscredential]::new('ignored', $SecureString)
+    $credential.GetNetworkCredential().Password
+}
+