+++ src/AvePoint.Elements/Private/Get-AvptModuleState.ps1
@@ -1,0 +1,14 @@
+function Get-AvptModuleState {
+    [CmdletBinding()]
+    param()
+
+    if (-not $script:AvptState) {
+        $script:AvptState = @{
+            Connection = $null
+            Token      = $null
+        }
+    }
+
+    $script:AvptState
+}
+